feat: filter and sort by context
This allows newly added tasks to be sorted in H, W, S, E order by context, and allows the user to pick a context to focus on via a hamburger menu. The behavior for marking a task as complete is changed so that the color of the context circle turns gray. This commit also adds an advance date button for testing, and fixes the corresponding removeCompletedTasks method.